fix: fix memory leak with store logs worker not closed
When calling the RunInTransaction store function, a sub
store is created everytime. When creating a new store,
a logs worker is created also using context.Background,
and the stop function was never called. This commit
fixes that by simply adding the logs worker closing
inside the ledger close and calling the close when we
finish the transaction